- add new SharedKeyCommands to ownCloudAppShared.framework, to implement key commands that should be available in app and extensions - move PasscodeViewController key commands over to SharedKeyCommands
